21.댓글 삭제

+++ src/main/java/shop/tenco/blog/user/UserRepository.java
@@ -48,7 +48,7 @@ public class UserRepository {
     
     
     @Transactional 
-    public void update(String password, Long userId){
+    public void update(String password, int userId){
         Query query = em.createNativeQuery("update user_tb set password = ? where id = ? ");
         query.setParameter(1, password);
         query.setParameter(2, userId);
